Add flex to body tag and change values in footer
The body tag was given the properties of display: flex and a flex
direction of column. The footer had it's flex values changed to
flex: 1, justify-content: space-between and align-items: end. Both of
these css changes were to move the content of the footer to the bottom
of the screen for the user.