Remove flex from footer and add value to body
Removed the flex property from the footer class and added
justify-content: space-between to the body class. This allows for the
footer to be in the same part of the screen but for its backround not
to take up most of the screen and moves the logo, buttons and input to
the middle of the screen vertically.